修复完善删除文件移动到回收站

+++ application/index/controller/Recycle.php
@@ -5,6 +5,8 @@ namespace app\index\controller;
 use app\common\controller\Home;
 use app\common\model\FileManage;
 use app\common\model\Folders;
+use app\common\model\Profit;
+use app\common\model\Shares;
 use app\common\model\Stores;
 
 class Recycle extends Home
@@ -37,6 +39,11 @@ class Recycle extends Home
             foreach ($stores as $file){
                 // 删除文件
                 $file->delete(true);
+                // 删除分享信息
+                Profit::where('file_id',$file['id'])
+                    ->where('uid',$uid)
+                    ->delete();
+
                 $success_ids[] = $file['id'];
             }
 
@@ -44,6 +51,24 @@ class Recycle extends Home
             foreach ($folders as $dir){
                 // 删除文件
                 $dir->delete(true);
+                // 获取子文件夹
+                $folder_object = getFolderChildFiles($uid,$dir['id']);
+
+                // 子文件夹
+                if(!empty($folder_object['folder'])){
+                    Folders::where('id','in',$folder_object['folder'])->delete(true);
+                }
+
+                // 子文件
+                if(!empty($folder_object['file'])){
+                    // 删除文件
+                    Stores::where('id','in',$folder_object['file'])->delete(true);
+                    // 删除分享信息
+                    Profit::where('file_id','in',$folder_object['file'])
+                        ->where('uid',$uid)
+                        ->delete();
+                }
+
                 $success_ids[] = $dir['id'];
             }
 
@@ -114,6 +139,12 @@ class Recycle extends Home
                 foreach ($stores as $file){
                     // 删除文件
                     $file->delete(true);
+
+                    // 删除分享信息
+                    Profit::where('file_id',$file['id'])
+                        ->where('uid',$uid)
+                        ->delete();
+
                     $success_ids[] = $file['id'];
                 }
             }catch (\Throwable $e){
@@ -127,6 +158,25 @@ class Recycle extends Home
                 foreach ($folders as $dir){
                     // 删除文件
                     $dir->delete(true);
+
+                    // 获取子文件夹
+                    $folder_object = getFolderChildFiles($uid,$dir['id']);
+
+                    // 子文件夹
+                    if(!empty($folder_object['folder'])){
+                        Folders::where('id','in',$folder_object['folder'])->delete(true);
+                    }
+
+                    // 子文件
+                    if(!empty($folder_object['file'])){
+                        // 删除文件
+                        Stores::where('id','in',$folder_object['file'])->delete(true);
+                        // 删除分享信息
+                        Profit::where('file_id','in',$folder_object['file'])
+                            ->where('uid',$uid)
+                            ->delete();
+                    }
+
                     $success_ids[] = $dir['id'];
                 }
             }catch (\Throwable $e){
